# swagger.yaml
openapi: 3.0.0
info:
  title: BiteBuddies API
  description: API documentation for BiteBuddies application services
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Account Service
  - url: http://localhost:5001
    description: Availability Service
  - url: http://localhost:5002
    description: Restaurant Service
  - url: http://localhost:5006
    description: Transcription Service
  - url: http://localhost:5007
    description: Chatbot Service
  - url: http://localhost:5010
    description: Search Service
...
paths:
...
  /test-db:
    get:
      tags:
        - Availability
      summary: Test database connection
      description: Check if the MongoDB connection is working
      responses:
        "200":
          description: Database connection is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "healthy"
                  message:
                    type: string
                    example: "Database is working"
        "500":
          description: Database connection error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: "Database connection error"
  /availability/{user_email}:
    get:
      tags:
        - Availability
      summary: Get all availability slots
      description: Retrieve all availability slots for a user
      parameters:
        - name: user_email
          in: path
          required: true
          description: User's email address
          schema:
            type: string
      responses:
        "200":
          description: List of availability slots
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Availability"
        "500":
          description: Database connection error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
  /availability/{user_email}/{date}:
    get:
      tags:
        - Availability
      summary: Get availability by date
      description: Retrieve availability slots for a user on a specific date
      parameters:
        - name: user_email
          in: path
          required: true
          description: User's email address
          schema:
            type: string
        - name: date
          in: path
          required: true
          description: Date in YYYY-MM-DD format
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of availability slots for the specified date
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Availability"
        "500":
          description: Database connection error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
  /availability/dates/{user_email}:
    get:
      tags:
        - Availability
      summary: Get available dates
      description: Retrieve all dates for which a user has availability slots
      parameters:
        - name: user_email
          in: path
          required: true
          description: User's email address
          schema:
            type: string
      responses:
        "200":
          description: List of dates with availability slots
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    type: object
                    properties:
                      dates:
                        type: array
                        items:
                          type: string
                          format: date
        "500":
          description: Database connection error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
  /availability:
    post:
      tags:
        - Availability
      summary: Create availability slot
      description: Create a new availability slot for a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_email
                - date
                - start_time
                - end_time
                - restaurant
              properties:
                user_email:
                  type: string
                  description: User's email address
                date:
                  type: string
                  format: date
                  description: Date in YYYY-MM-DD format
                start_time:
                  type: string
                  description: Start time in HH:MM format
                end_time:
                  type: string
                  description: End time in HH:MM format
                restaurant:
                  type: string
                  description: Restaurant identifier
                status:
                  type: string
                  description: Availability status
                  default: available
      responses:
        "201":
          description: Availability slot created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 201
                  data:
                    $ref: "#/components/schemas/Availability"
        "400":
          description: Bad request - missing required fields or invalid date/time format
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 400
                  message:
                    type: string
        "500":
          description: Database connection error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
  /availability/delete:
    post:
      tags:
        - Availability
      summary: Delete availability slot
      description: Delete an availability slot for a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_email
                - date
                - start_time
              properties:
                user_email:
                  type: string
                  description: User's email address
                date:
                  type: string
                  format: date
                  description: Date in YYYY-MM-DD format
                start_time:
                  type: string
                  description: Start time in HH:MM format
      responses:
        "200":
          description: Slot deleted or no matching slot found
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
        "400":
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 400
                  message:
                    type: string
        "500":
          description: Database connection error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
  /availability/search:
    post:
      tags:
        - Availability
      summary: Search availability slots
      description: Search for availability slots matching specific criteria
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - start_time
                - end_time
                - date
                - restaurant
                - status
              properties:
                start_time:
                  type: string
                  description: Start time in HH:MM format
                end_time:
                  type: string
                  description: End time in HH:MM format
                date:
                  type: string
                  format: date
                  description: Date in YYYY-MM-DD format
                restaurant:
                  type: string
                  description: Restaurant identifier
                status:
                  type: string
                  description: Availability status
      responses:
        "200":
          description: List of matching availability slots
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    type: object
                    properties:
                      matches:
                        type: array
                        items:
                          $ref: "#/components/schemas/Availability"
        "400":
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 400
                  message:
                    type: string
        "500":
          description: Database connection error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
  /availability/check:
    post:
      tags:
        - Availability
      summary: Check availability
      description: Check if a user is available at a specific time and location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - date
                - startTime
                - endTime
                - user_email
                - restaurant
                - status
              properties:
                date:
                  type: string
                  format: date
                  description: Date in YYYY-MM-DD format
                startTime:
                  type: string
                  description: Start time in HH:MM format
                endTime:
                  type: string
                  description: End time in HH:MM format
                user_email:
                  type: string
                  description: User's email address
                restaurant:
                  type: string
                  description: Restaurant identifier
                status:
                  type: string
                  description: Availability status to check for
      responses:
        "200":
          description: Availability check result
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  available:
                    type: boolean
                  data:
                    type: object
        "400":
          description: Bad request - missing fields or invalid date/time format
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 400
                  message:
                    type: string
        "500":
          description: Database connection error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
  /availability/update_status:
    post:
      tags:
        - Availability
      summary: Update availability status
      description: Update the status of an availability slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_email
                - date
                - startTime
                - endTime
                - restaurant
                - status
              properties:
                user_email:
                  type: string
                  description: User's email address
                date:
                  type: string
                  format: date
                  description: Date in YYYY-MM-DD format
                startTime:
                  type: string
                  description: Start time in HH:MM format
                endTime:
                  type: string
                  description: End time in HH:MM format
                restaurant:
                  type: string
                  description: Restaurant identifier
                status:
                  type: string
                  description: New availability status
      responses:
        "200":
          description: Status updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/Availability"
        "400":
          description: Bad request - missing fields, invalid date/time format, or slot not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 400
                  message:
                    type: string
        "500":
          description: Database connection error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string

  # =========== RESTAURANT SERVICE ===========
...
components:
  schemas:
...
    Availability:
      type: object
      properties:
        id:
          type: string
          description: Availability slot unique identifier
        user_email:
          type: string
          description: User's email address
        date:
          type: string
          format: date
          description: Date in YYYY-MM-DD format
        start_time:
          type: string
          description: Start time in HH:MM format
        end_time:
          type: string
          description: End time in HH:MM format
        restaurant:
          type: string
          description: Restaurant identifier
        status:
          type: string
          description: Availability status (e.g., 'available', 'booked')
          default: available
...